Refactor psbt module into submodules
Since there are already 2 independent PSBT extensions, and more
will come, it's cleaner to have each extension in its own file.
